#18 Refactored router file structure
