+++ routes/api/orders.js
@@ -1,3 +1,6 @@
+var express = require('express');
+var router = express.Router();
+
 var orders = {};
 
 //This will be my mock database for 'searching', will be repleaced with the actual database
@@ -162,8 +165,28 @@ var inflateResponseObject = function(data,billingInfo,shippingInfo,customerInfo,
 	return order;
 };
 
-orders.get = function(req,res,next,orderId,billingInfo,shippingInfo,customerInfo,items){
+router.get('/', function(req,res,next){
 	var orders = [];
+	
+	//int : Id of the order
+	var orderId = req.param('orderId');
+	//bool : if the billing info should be returned
+	var billingInfo = req.param('billingInfo');
+	billingInfo = billingInfo == 'true';
+	//bool : if the shipping info should be returned
+	var shippingInfo = req.param('shippingInfo');
+	shippingInfo = shippingInfo == 'true';
+	//bool : if the customer info should be returned
+	var customerInfo = req.param('customerInfo');
+	customerInfo = customerInfo == 'true';
+	//bool : if the list of items should be returned
+	var items = req.param('items');	
+	items = items == 'true';
+	
+	if (!orderId) {
+		res.status(400).send('400 Bad request: orderId is needed');
+	}
+	
 	database.forEach(function(data){
 		if(data['id'] == orderId){
 			var order = inflateResponseObject(data,billingInfo,shippingInfo,customerInfo,items);
@@ -174,10 +197,30 @@ orders.get = function(req,res,next,orderId,billingInfo,shippingInfo,customerInfo
 		"orders": orders
 	};
 	res.json(responseObj);
-};
+});
 
-orders.query = function(req,res,next,address,billingAddress,customerId,billingInfo,shippingInfo,customerInfo,shippingInfo,customerInfo,items){
+router.get('/search', function(req,res,next){
 	var orders = [];
+		//String : Shipping address
+	var address = req.param('address');
+	//bool : False by default, True will search by billing
+	var billingAddress = req.param('billingAddress');
+	billingAddress = billingAddress == 'true';
+	//int : Customer Id to help refine search. Optional
+	var customerId = req.param('customerId');
+	//bool : if the billing info should be returned
+        var billingInfo = req.param('billingInfo');
+	billingInfo = billingInfo == 'true';
+        //bool : if the shipping info should be returned
+        var shippingInfo = req.param('shippingInfo');
+	shippingInfo = shippingInfo == 'true'; 
+        //bool : if the customer info should be returned
+        var customerInfo = req.param('customerInfo');
+	customerInfo = customerInfo == 'true';
+        //bool : if the list of items should be returned
+        var items = req.param('items');
+	items = items == 'true';
+	
 	database.forEach(function(data){
 		var found = false;
 		if(billingAddress == true && (address && address.length > 0)){
@@ -210,6 +253,6 @@ orders.query = function(req,res,next,address,billingAddress,customerId,billingIn
 		"orders": orders
 	};
 	res.json(responseObj);
-};
+});
 
-module.exports = orders;
+module.exports = router;
